requirements.txt
certifi==2021.5.30
click==8.0.1
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.0.1
keras==2.7.0
tensorflow-cpu==2.7.0
MarkupSafe==2.0.1
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
boto3==1.20.0
Pillow==8.4.0